chore: split the WFS query compiler into focused modules
