+++ src/helpers/wfs_engine/properties.ts
@@ -1,0 +1,97 @@
+/**
+ * Property resolution and validation helpers for the structured WFS query compiler.
+ *
+ * This module centralizes:
+ * - geometry property lookup
+ * - property existence checks
+ * - non-geometry validation for select/order/filter compilation
+ */
+
+import type { Collection, CollectionProperty } from "@ignfab/gpf-schema-store";
+
+/**
+ * Lists available property names for a feature type, mainly for error reporting.
+ *
+ * @param featureType Feature type definition loaded from the embedded catalog.
+ * @returns A comma-separated list of property names.
+ */
+function getPropertyList(featureType: Collection) {
+  return featureType.properties.map((property) => property.name).join(", ");
+}
+
+/**
+ * Returns every geometry-like property exposed by a feature type.
+ *
+ * @param featureType Feature type definition loaded from the embedded catalog.
+ * @returns The list of properties carrying a `defaultCrs`.
+ */
+function getGeometryProperties(featureType: Collection) {
+  return featureType.properties.filter((property) => property.defaultCrs);
+}
+
+/**
+ * Resolves the single geometry property expected by the query compiler.
+ *
+ * @param featureType Feature type definition loaded from the embedded catalog.
+ * @returns The unique geometry property for the feature type.
+ */
+export function getGeometryProperty(featureType: Collection) {
+  const geometryProperties = getGeometryProperties(featureType);
+  if (geometryProperties.length === 0) {
+    throw new Error(`Le type '${featureType.id}' n'expose aucune propriété géométrique exploitable dans le catalogue embarqué.`);
+  }
+  if (geometryProperties.length > 1) {
+    throw new Error(`Le type '${featureType.id}' expose plusieurs propriétés géométriques dans le catalogue embarqué : ${geometryProperties.map((property) => property.name).join(", ")}.`);
+  }
+  return geometryProperties[0];
+}
+
+/**
+ * Loads a property by exact name and throws a descriptive error when it does not exist.
+ *
+ * @param featureType Feature type definition loaded from the embedded catalog.
+ * @param propertyName Exact property name requested by the caller.
+ * @returns The matching property metadata.
+ */
+function getPropertyOrThrow(featureType: Collection, propertyName: string) {
+  const property = featureType.properties.find((candidate) => candidate.name === propertyName);
+  if (!property) {
+    //throw new Error(`La propriété '${propertyName}' n'existe pas pour '${featureType.id}'. Utiliser une propriété parmi : ${getPropertyList(featureType)}.`);
+    throw new Error(`La propriété '${propertyName}' n'existe pas pour '${featureType.id}'. Appelle \`gpf_wfs_describe_type\` pour obtenir la liste des propriétés disponibles.`);
+  }
+  return property;
+}
+
+/**
+ * Ensures that a property exists and is not the geometry column of the feature type.
+ *
+ * @param featureType Feature type definition loaded from the embedded catalog.
+ * @param geometryProperty Geometry property already resolved for the feature type.
+ * @param propertyName Exact property name requested by the caller.
+ * @param message Error message template used when the property is geometric.
+ * @returns The matching non-geometric property metadata.
+ */
+export function ensureNonGeometryProperty(featureType: Collection, geometryProperty: CollectionProperty, propertyName: string, message: string) {
+  const property = getPropertyOrThrow(featureType, propertyName);
+  if (property.name === geometryProperty.name || property.defaultCrs) {
+    throw new Error(message.replace("{property}", property.name));
+  }
+  return property;
+}
+
+/**
+ * Validates a selected property name and returns the exact property name to expose.
+ *
+ * @param featureType Feature type definition loaded from the embedded catalog.
+ * @param geometryProperty Geometry property already resolved for the feature type.
+ * @param propertyName Raw selected property name.
+ * @returns The validated non-geometric property name.
+ */
+export function compileSelectProperty(featureType: Collection, geometryProperty: CollectionProperty, propertyName: string) {
+  return ensureNonGeometryProperty(
+    featureType,
+    geometryProperty,
+    propertyName,
+    "La propriété '{property}' est géométrique. `select` accepte uniquement des propriétés non géométriques."
+  ).name;
+}
